Fix overflow in DataLoc positions
Multiple values in the DataLoc improperly convert from long to int,
which could lead to overflows when bit positions grow to larger than
Int.MaxValue. This leads to assertions on DataLoc calculations that are
expeced to only be positive. This fixes the calculations so we keep bit
positions as a long type to avoid these overflows. Also removes some
unnecessary math.max calls since we know certain calculations will
always be positive.

DAFFODIL-2785

+++ daffodil-cli/src/main/scala/org/apache/daffodil/cli/Main.scala
@@ -829,7 +829,7 @@ object Main {
                     // not streaming mode, and there is more data available,
                     // so show left over data warning
                     val Dump = new DataDumper
-                    val bitsAlreadyConsumed = loc.bitPos0b % 8
+                    val bitsAlreadyConsumed = (loc.bitPos0b % 8).toInt
                     val firstByteString = if (bitsAlreadyConsumed != 0) {
                       val bitsToDisplay = 8 - bitsAlreadyConsumed
                       val pbp = inStream.inputSource.position + 1
